#43 Updated Transaction's ListTile, moved files, etc.
- Added the exchangedAmount/Currency to FundDetails
- Transaction's ListTile now show the converted amount and currency

- Moved CreateTxnForm and EditTxnForm to a seperate debug_form folder
